tools/carla/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vsbs-carla"
version = "0.1.0"
description = "CARLA bridge for the VSBS autonomous-service-booking demo loop."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "Divya Mohan", email = "contact@dmj.one" }
]
keywords = ["carla", "vsbs", "autonomous-vehicle", "phm", "service-booking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26",
    "pydantic>=2.5",
    "httpx>=0.27",
    "httpx-sse>=0.4",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "pyyaml>=6",
]

[project.optional-dependencies]
carla = ["carla==0.10.0"]
dev = ["pytest>=7.4", "pytest-asyncio>=0.23"]

[project.scripts]
vsbs-carla-demo = "vsbs_carla.scripts.run_demo:main"
vsbs-carla-record = "vsbs_carla.scripts.record_trace:main"

[tool.setuptools.packages.find]
include = ["vsbs_carla*"]
exclude = ["tests*", "replay*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
